Rename P2P errors file to exceptions
