+++ elements/lisk-p2p/src/index.ts
@@ -13,6 +13,6 @@
  *
  */
 export * from './disconnect_status_codes';
-export * from './errors';
+export * from './exceptions';
 export * from './events';
 export * from './p2p';
