+++ elements/lisk-p2p/src/validation.ts
@@ -19,7 +19,7 @@ import {
 	InvalidProtocolMessageError,
 	InvalidRPCRequestError,
 	InvalidRPCResponseError,
-} from './errors';
+} from './exceptions';
 
 import {
 	INCOMPATIBLE_NETWORK_REASON,
